Update BepInEx.PluginInfoProps.props for better Inter-Mod Referencing
In version 1.0.0 `MyPluginInfo`, formerly `PluginInfo`, was an internal class. However due to the v1.0.0 `PluginInfo` not having a project-specific namespace, which came the next update when the internal class was converted to a public class as well, it's properties were not accessible to the plugin's scope. 
With the update seen in v1.1.0 to correctly namespace the class, it can be kept as `internal` and still allow it's variables to be accessed by the defining mod.

## Benefits
When a library mod is referenced by another plugin, the receiving plugin can scope into the library's namespace and also use BepInEx.PluginInfoProps without any confusion about which `MyPluginInfo` is being referenced.