Add Template CI Github Workflow (#13)
This PR adds a github workflow that will attempt to create projects from the templates in all supported dotnet SDK versions (6, 7, & 8) when a push to any branch or a PR is made (if the action is approved to run).

This acts as two checks. The first, to make sure templates can be used at all. The second, to make sure templates can be used in all of the supported SDK versions.